move permissions checks back to headless detector
